requirements.txt
chardet~=3.0.4
django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.10.0
djongo==1.3.6
dnspython==1.16.0
jsonfield==3.1.0
numpy==1.19.5
pandas==1.3.3
pymongo~=3.12.0
drf-yasg==1.20.0
requests~=2.22.0
beautifulsoup4~=4.10.0
openpyxl~=3.0.9
unidecode~=1.3.2